ADMIN CONTROLLER FINI

+++ boutique/controller/AdminController.php
@@ -41,6 +41,10 @@ class AdminController
 
     public function suprimmerProduit()
     {
+        if (!isset($_GET['id'])) {
+            header("Location: index.php?route=admin");
+            exit();
+        }
         $id = (int)$_GET['id'];
         $this->repo->delete($id);
         header("Location: index.php?route=admin");
